requirements.txt
flask==3.0.0
sqlalchemy==2.0.23
werkzeug==3.0.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
pywebpush==1.14.0
cryptography==41.0.7
cloudinary==1.38.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0